@@ -15,7 +15,7 @@ long_description = file: README.md
 long_description_content_type = text/markdown
 
 [options]
-python_requires = >=3.6
+python_requires = >=3.9
 install_requires =
     requests
     python-dateutil
