@@ -7,11 +7,10 @@
 
 from __future__ import annotations
 
-from datetime import datetime
 from typing import Any, Generator, List
 
 from c8y_api.model._base import CumulocityResource
-from c8y_api.model._util import _DateUtil, _QueryUtil
+from c8y_api.model._util import _QueryUtil
 from c8y_api.model.managedobjects import ManagedObjectUtil, ManagedObject, Device, Availability, DeviceGroup
 
 
@@ -37,14 +36,25 @@ class Inventory(CumulocityResource):
              A ManagedObject instance
 
         Raises:
-            KeyError if the ID is not defined within the database
+            KeyError:  if the ID is not defined within the database
         """
         managed_object = ManagedObject.from_json(self._get_object(id))
         managed_object.c8y = self.c8y  # inject c8y connection into instance
         return managed_object
 
-    def get_all(self, type: str = None, fragment: str = None, name: str = None, owner: str = None,  # noqa (type)
-                limit: int = None, page_size: int = 1000):
+    def get_all(
+            self,
+            expression: str = None,
+            type: str = None,
+            fragment: str = None,
+            name: str = None,
+            owner: str = None,
+            query: str = None,
+            text: str = None,
+            ids: List[str | int] = None,
+            limit: int = None,
+            page_size: int = 1000
+    ) -> List[ManagedObject]:
         """ Query the database for managed objects and return the results
         as list.
 
@@ -54,10 +64,62 @@ class Inventory(CumulocityResource):
         Returns:
             List of ManagedObject instances
         """
-        return list(self.select(type=type, fragment=fragment, name=name, owner=owner, limit=limit, page_size=page_size))
+        return list(self.select(
+            expression=expression,
+            type=type,
+            fragment=fragment,
+            name=name,
+            owner=owner,
+            query=query,
+            text=text,
+            ids=ids,
+            limit=limit,
+            page_size=page_size))
 
-    def select(self, type: str = None, fragment: str = None, name: str = None, owner: str = None,  # noqa (type)
-               query: str = None, limit: int = None, page_size: int = 1000) -> Generator[ManagedObject]:
+    def get_count(
+            self,
+            expression: str = None,
+            type: str = None,
+            fragment: str = None,
+            name: str = None,
+            owner: str = None,
+            query: str = None,
+            text: str = None,
+            ids: List[str | int] = None
+    ) -> int:
+        """Calculate the number of potential results of a database query.
+
+        This function uses the same parameters as the `select` function.
+
+        Returns:
+            Number of potential results
+        """
+        base_query = self._prepare_query(
+            expression=expression,
+            type=type,
+            fragment=fragment,
+            name=name,
+            owner=owner,
+            query=query,
+            text=text,
+            ids=ids,
+            page_size=1)
+        return self._get_count(base_query)
+
+    def select(
+            self,
+            expression: str = None,
+            type: str = None,
+            fragment: str = None,
+            name: str = None,
+            owner: str = None,
+            query: str = None,
+            text: str = None,
+            ids: List[str|int] = None,
+            limit: int = None,
+            page_size: int = 1000,
+            page_number: int = None
+    ) -> Generator[ManagedObject]:
         """ Query the database for managed objects and iterate over the
         results.
 
@@ -69,6 +131,9 @@ class Inventory(CumulocityResource):
         combined (within reason).
 
         Args:
+            expression (str):  Arbitrary filter expression which will be
+                passed to Cumulocity without change; all other filters
+                are ignored if this is provided
             type (str):  Managed object type
             fragment (str):  Name of a present custom/standard fragment
             name (str):  Name of the managed object
@@ -78,28 +143,66 @@ class Inventory(CumulocityResource):
             owner (str):  Username of the object owner
             query (str):  Complex query to execute; all other filters are
                 ignored if such a custom query is provided
+            text (str): Text value of any object property.
+            ids (List[str|int]): Specific object ID to select.
             limit (int): Limit the number of results to this number.
             page_size (int): Define the number of events which are read (and
                 parsed in one chunk). This is a performance related setting.
+            page_number (int): Pull a specific page; this effectively disables
+                automatic follow-up page retrieval.
 
         Returns:
             Generator for ManagedObject instances
         """
-        return self._select(ManagedObject.from_json, type=type, fragment=fragment, name=name, owner=owner,
-                            query=None, limit=limit, page_size=page_size)
+        return self._select(
+            ManagedObject.from_json,
+            expression=expression,
+            type=type,
+            fragment=fragment,
+            name=name,
+            owner=owner,
+            query=query,
+            text=text,
+            ids=ids,
+            limit=limit,
+            page_size=page_size,
+            page_number=page_number)
 
-    def _select(self, jsonyfy_func, type: str = None, fragment: str = None, name: str = None,  # noqa
-                owner: str = None, query: str = None, limit: int = None, page_size: int = 1000) -> Generator[Any]:
+    @classmethod
+    def _prepare_query_param(cls, query, filters):
+        """Potentially extend a query parameter with additional filters.
 
-        query_filters = []
+        If there are no additional filters, the query parameter is not
+        touched. Otherwise, a complete query is prepared which consists of
+        the original query plus all additional filters.
+        """
+        if not filters:
+            return query
+        add_filters = ' and '.join(filters)
+        if not query:
+            return add_filters
+        return query + ' $filter=(' + add_filters + ')'
 
+    def _prepare_query(
+            self,
+            expression: str = None,
+            type: str = None,
+            fragment: str = None,
+            name: str = None,
+            owner: str = None,
+            query: str = None,
+            **kwargs
+    ) -> str:
+        """The inventory API features a query API that needs some additional
+        preparations before we can actually invoke the queries."""
+        if expression:
+            return self._build_base_query(expression=expression)
         # if there is no custom query, we check whether standard filters need to
         # be translated into a query
         if not query and name:
-
             # A name filter can only be expressed as a query, which then
             # triggers "query mode" (all filters are translated into a query)
-            query_filters.append(f"name eq '{_QueryUtil.encode_odata_query_value(name)}'")
+            query_filters = [f"name eq '{_QueryUtil.encode_odata_query_value(name)}'"]
 
             if type:
                 query_filters.append(f"type eq '{type}'")
@@ -107,37 +210,38 @@ class Inventory(CumulocityResource):
                 query_filters.append(f"owner eq '{owner}'")
             if fragment:
                 query_filters.append(f"has({fragment})")
-            if len(query_filters) == 1:
-                query = query_filters[0]
-            else:
-                query = '$filter=(' + ' and '.join(query_filters) + ')'
+            query = ' and '.join(query_filters)
 
         if query:
-            base_query = self._build_base_query(query=query, page_size=page_size)
-        else:
-            base_query = self._build_base_query(type=type, fragment=fragment, owner=owner, page_size=page_size)
+            # all parameters except page_size (which is not a filter) are ignored
+            return self._build_base_query(query=query, page_size=kwargs.get('page_size', None))
+        return self._build_base_query(type=type, fragment=fragment, owner=owner, **kwargs)
 
-        return super()._iterate(base_query, limit, jsonyfy_func)
+    def _select(self, jsonify_func, **kwargs) -> Generator[Any]:
+        """Generic select function to be used by derived classes as well."""
+        page_number = kwargs.pop('page_number', None)
+        limit = kwargs.pop('limit', None)
+        return super()._iterate(self._prepare_query(**kwargs), page_number, limit, jsonify_func)
 
     def create(self, *objects: ManagedObject):
         """Create managed objects within the database.
 
         Args:
-            objects (*ManagedObject): collection of ManagedObject instances
+           *objects (ManagedObject): collection of ManagedObject instances
         """
         super()._create(ManagedObject.to_json, *objects)
 
     def update(self, *objects: ManagedObject):
-        """ Write changes to the database.
+        """Write changes to the database.
 
         Args:
-            objects (*ManagedObject): collection of ManagedObject instances
+           *objects (ManagedObject): collection of ManagedObject instances
 
-        See also function ManagedObject.update which parses the result.
+        See also function `ManagedObject.update` which parses the result.
         """
         super()._update(ManagedObject.to_diff_json, *objects)
 
-    def apply_to(self, object_model: ManagedObject, *object_ids):
+    def apply_to(self, object_model: ManagedObject | dict, *object_ids):
         """Apply a change to multiple already existing objects.
 
         Applies the details of a model object to a set of already existing
@@ -146,9 +250,10 @@ class Inventory(CumulocityResource):
         Note: This will take the full details, not just the updates.
 
         Args:
-            object_model (ManagedObject): ManagedObject instance holding
-                the change structure (e.g. a specific fragment)
-            object_ids (*str): a collection of ID of already existing
+            object_model (ManagedObject|dict): ManagedObject instance holding
+                the change structure (e.g. a specific fragment) or simply a
+                dictionary representing the diff JSON.
+            *object_ids (str): a collection of ID of already existing
                 managed objects within the database
         """
         super()._apply_to(ManagedObject.to_full_json, object_model, *object_ids)
@@ -191,6 +296,7 @@ class Inventory(CumulocityResource):
         result_json = self.c8y.get(self.build_object_path(mo_id) + '/' + ManagedObject.Resource.SUPPORTED_SERIES)
         return result_json[ManagedObject.Fragment.SUPPORTED_SERIES]
 
+
 class DeviceInventory(Inventory):
     """Provides access to the Device Inventory API.
 
@@ -228,15 +334,34 @@ class DeviceInventory(Inventory):
             A Device instance
 
         Raises:
-            KeyError if the ID is not defined within the database
+            KeyError:  if the ID is not defined within the database
         """
         device = Device.from_json(self._get_object(id))
         device.c8y = self.c8y
         return device
 
-    def select(self, type: str = None, name: str = None, owner: str = None,  # noqa (type, args)
-               query: str = None, limit: int = None, page_size: int = 100) -> Generator[Device]:
-        # pylint: disable=arguments-differ
+    @classmethod
+    def _prepare_device_query_param(cls, query: str) -> str:
+        if query:
+            # insert after opening bracket or at the beginning
+            insert_at = query.find('filter=', ) + 1
+            query = query[:insert_at] + "has(c8y_IsDevice) and " + query[insert_at:]
+        return query
+
+    def select(  # noqa (order)
+            self,
+            expression: str = None,
+            type: str = None,
+            name: str = None,
+            owner: str = None,
+            query: str = None,
+            text: str = None,
+            ids: List[str | int] = None,
+            limit: int = None,
+            page_size: int = 100,
+            page_number: int = None
+    ) -> Generator[Device]:
+        # pylint: disable=arguments-differ, arguments-renamed
         """ Query the database for devices and iterate over the results.
 
         This function is implemented in a lazy fashion - results will only be
@@ -246,7 +371,13 @@ class DeviceInventory(Inventory):
         to objects which meet the filters specification.  Filters can be
         combined (within reason).
 
+        Note: this variant doesn't allow filtering by fragment because the
+        `c8y_IsDevice` fragment is automatically filtered.
+
         Args:
+            expression (str):  Arbitrary filter expression which will be
+                passed to Cumulocity without change; all other filters
+                are ignored if this is provided
             type (str):  Device type
             name (str):  Name of the device
                 Note: The Cumulocity REST API does not support filtering for
@@ -255,19 +386,44 @@ class DeviceInventory(Inventory):
             owner (str):  Username of the object owner
             query (str):  Complex query to execute; all other filters are
                 ignored if such a custom query is provided
+            text (str): Text value of any object property.
+            ids (List[str|int]): Specific object ID to select.
             limit (int): Limit the number of results to this number.
             page_size (int): Define the number of events which are read (and
                 parsed in one chunk). This is a performance related setting.
+            page_number (int): Pull a specific page; this effectively disables
+                automatic follow-up page retrieval.
 
         Returns:
             Generator for Device objects
         """
-        return self._select(ManagedObject.from_json, type=type, fragment='c8y_IsDevice', name=name, owner=owner,
-                            query=query, limit=limit, page_size=page_size)
+        return super()._select(
+            Device.from_json,
+            type=type,
+            fragment='c8y_IsDevice',
+            name=name,
+            owner=owner,
+            query=self._prepare_device_query_param(query),
+            text=text,
+            ids=ids,
+            limit=limit,
+            page_size=page_size,
+            page_number=page_number)
 
-    def get_all(self, type: str = None, name: str = None, owner: str = None,   # noqa (type, parameters)
-                page_size: int = 100) -> List[Device]:
-        # pylint: disable=arguments-differ
+    def get_all(  # noqa (changed signature)
+            self,
+            expression: str = None,
+            type: str = None,
+            name: str = None,
+            owner: str = None,
+            query: str = None,
+            text: str = None,
+            ids: List[str | int] = None,
+            limit: int = None,
+            page_size: int = 100,
+            page_number: int = None
+    ) -> List[Device]:
+        # pylint: disable=arguments-differ, arguments-renamed
         """ Query the database for devices and return the results as list.
 
         This function is a greedy version of the `select` function. All
@@ -276,12 +432,49 @@ class DeviceInventory(Inventory):
         Returns:
             List of Device objects
         """
-        return list(self.select(type=type, name=name, owner=owner, page_size=page_size))
+        return list(self.select(
+            expression=expression,
+            type=type,
+            name=name,
+            owner=owner,
+            query=self._prepare_device_query_param(query),
+            text=text,
+            ids=ids,
+            limit=limit,
+            page_size=page_size,
+            page_number=page_number))
+
+    def get_count(  # noqa (changed signature)
+            self,
+            type: str = None,
+            name: str = None,
+            owner: str = None,
+            query: str = None,
+            text: str = None,
+            ids: List[str | int] = None
+    ) -> int:
+        # pylint: disable=arguments-differ, arguments-renamed
+        """Calculate the number of potential results of a database query.
+
+        This function uses the same parameters as the `select` function.
+
+        Returns:
+            Number of potential results
+        """
+        return self._get_count(super()._prepare_query(
+            type=type,
+            fragment='c8y_IsDevice',
+            name=name,
+            owner=owner,
+            query=self._prepare_device_query_param(query),
+            text=text,
+            ids=ids,
+            page_size=1))
 
     def delete(self, *devices: Device):
         """ Delete one or more devices and the corresponding within the database.
 
-        The objects can be specified as instances of an database object
+        The objects can be specified as instances of a database object
         (then, the id field needs to be defined) or simply as ID (integers
         or strings).
 
@@ -290,7 +483,7 @@ class DeviceInventory(Inventory):
         user from database.
 
         Args:
-            devices (*Device): Device objects within the database specified
+           *devices (Device): Device objects within the database specified
                 (with defined ID).
         """
         for d in devices:
@@ -306,20 +499,77 @@ class DeviceGroupInventory(Inventory):
     See also: https://cumulocity.com/api/#tag/Inventory-API
     """
 
-    def get(self, group_id):
+    def get(self, group_id: str):
         # pylint: disable=arguments-differ, arguments-renamed
         """ Retrieve a specific device group object.
 
-        :param group_id:  ID of the device group object
-        :return:  a DeviceGroup instance
-        :raises:  KeyError if the ID is not defined within the database
+        Args:
+            group_id (str):  ID of the device group object.
+
+        Returns:
+            DeviceGroup instance.
+
+        Raises:
+            KeyError:  if the ID is not defined within the database.
         """
         group = DeviceGroup.from_json(self._get_object(group_id))
         group.c8y = self.c8y
         return group
 
-    def select(self, type: str = DeviceGroup.ROOT_TYPE, parent: str | int = None, fragment: str = None,  # noqa
-               name: str = None, owner: str = None, query: str = None, page_size: int = 100) -> Generator[DeviceGroup]:
+    def _prepare_device_group_query(
+            self,
+            type: str,
+            parent: str | int = None,
+            fragment: str = None,
+            name: str = None,
+            owner: str = None,
+            query: str = None,
+            **kwargs
+    ) -> str:
+        # pylint: disable=arguments-differ, arguments-renamed
+
+        query_filters = []
+        # Both name and parent filters can only be expressed as a query,
+        # which then triggers "query mode"
+        if name:
+            query_filters.append(f"name eq '{_QueryUtil.encode_odata_query_value(name)}'")
+        if parent:
+            query_filters.append(f"bygroupid({parent})")
+            type = DeviceGroup.CHILD_TYPE  # noqa
+
+        # if any query was defined, all filters must be put into the query
+        if query_filters:
+            if type:
+                query_filters.append(f"type eq '{type}'")
+            if owner:
+                query_filters.append(f"owner eq '{owner}'")
+            if fragment:
+                query_filters.append(f"has({fragment}")
+
+        if query_filters:
+            query = self._prepare_query_param(query, query_filters)
+
+        if query:
+            page_size = kwargs.get('page_size', None)
+            return self._build_base_query(query=query, page_size=page_size)
+
+        return self._build_base_query(type=type, fragment=fragment, owner=owner, **kwargs)
+
+    def select(  # noqa (changed signature)
+            self,
+            expression: str = None,
+            type: str = DeviceGroup.ROOT_TYPE,
+            parent: str|int = None,
+            fragment: str = None,
+            name: str = None,
+            owner: str = None,
+            query: str = None,
+            text: str = None,
+            ids: List[str|int] = None,
+            limit: int = None,
+            page_size: int = 100,
+            page_number: int = None
+    ) -> Generator[DeviceGroup]:
         # pylint: disable=arguments-differ, arguments-renamed
         """ Select device groups by various parameters.
 
@@ -327,10 +577,13 @@ class DeviceGroupInventory(Inventory):
         parsed and returned one by one.
 
         The type of all DeviceGroup objects is fixed 'c8y_DeviceGroup',
-        'c8y_DeviceSubGroup' if searching by `parent` respectively. Hence
+        'c8y_DeviceSubGroup' if searching by `parent` respectively. Hence,
         manual filtering by type is not supported.
 
         Args:
+            expression (str):  Arbitrary filter expression which will be
+                passed to Cumulocity without change; all other filters
+                are ignored if this is provided
             type (bool):  Filter for root or child groups respectively.
                 Note: If set to None, no type filter will be applied which
                 will match all kinds of managed objects. If you want to
@@ -348,45 +601,77 @@ class DeviceGroupInventory(Inventory):
             owner (str): Username of the group owner
             query (str):  Complex query to execute; all other filters are
                 ignored if such a custom query is provided
+            text (str): Text value of any object property.
+            ids (List[str|int]): Specific object ID to select
+            limit (int): Limit the number of results to this number.
             page_size (int): Define the number of events which are read (and
                 parsed in one chunk). This is a performance related setting.
+            page_number (int): Pull a specific page; this effectively disables
+                automatic follow-up page retrieval.
 
         Returns:
             Generator of DeviceGroup instances
         """
-        query_filters = []
+        base_query = self._prepare_device_group_query(
+            expression=expression,
+            type=type,
+            parent=parent,
+            fragment=fragment,
+            name=name,
+            owner=owner,
+            query=query,
+            text=text,
+            ids=ids,
+            page_size=page_size)
+        return super()._iterate(base_query, page_number, limit=limit, parse_func=DeviceGroup.from_json)
 
-        # if there is no custom query, we check whether standard filters need to
-        # be translated into a query
-        if not query:
+    def get_count(  # noqa (changed signature)
+            self,
+            expression: str = None,
+            type: str = DeviceGroup.ROOT_TYPE,
+            parent: str|int = None,
+            fragment: str = None,
+            name: str = None,
+            owner: str = None,
+            query: str = None,
+            text: str = None,
+            ids: List[str|int] = None,
+    ) -> int:
+        # pylint: disable=arguments-differ, arguments-renamed
+        """Calculate the number of potential results of a database query.
 
-            # Both name and parent filters can only be expressed as a query,
-            # which then triggers "query mode"
-            if name:
-                query_filters.append(f"name eq '{_QueryUtil.encode_odata_query_value(name)}'")
-            if parent:
-                query_filters.append(f"bygroupid({parent})")
-                type = DeviceGroup.CHILD_TYPE  # noqa
+        This function uses the same parameters as the `select` function.
 
-            # if any query was defined, all filters must be put into the query
-            if query_filters:
-                if type:
-                    query_filters.append(f"type eq '{type}'")
-                if owner:
-                    query_filters.append(f"owner eq '{owner}'")
-                if fragment:
-                    query_filters.append(f"has({fragment}")
-                query = '$filter=' + ' and '.join(query_filters)
+        Returns:
+            Number of potential results
+        """
+        base_query = self._prepare_device_group_query(
+            expression=expression,
+            type=type,
+            parent=parent,
+            fragment=fragment,
+            name=name,
+            owner=owner,
+            query=query,
+            text=text,
+            ids=ids,
+            page_size=1)
+        return self._get_count(base_query)
 
-        if query:
-            base_query = self._build_base_query(query=query, page_size=page_size)
-        else:
-            base_query = self._build_base_query(type=type, fragment=fragment, owner=owner, page_size=page_size)
-
-        return super()._iterate(base_query, limit=9999, parse_func=DeviceGroup.from_json)
-
-    def get_all(self, type: str = DeviceGroup.ROOT_TYPE, parent: str | int = None, fragment: str = None, # noqa
-                name: str = None, owner: str = None, page_size: int = 100):  # noqa
+    def get_all(  # noqa (changed signature)
+            self,
+            expression: str = None,
+            type: str = DeviceGroup.ROOT_TYPE,
+            parent: str | int = None,
+            fragment: str = None,
+            name: str = None,
+            owner: str = None,
+            query: str = None,
+            text: str = None,
+            ids: List[str|int] = None,
+            page_size: int = 100,
+            page_number: int = None
+    ) -> List[DeviceGroup]:
         # pylint: disable=arguments-differ, arguments-renamed
         """ Select managed objects by various parameters.
 
@@ -394,24 +679,36 @@ class DeviceGroupInventory(Inventory):
         collect the entire result set before returning.
 
         Returns:
-            List of DeviceGroup instances
+            List of DeviceGroup instances.
         """
-        return list(self.select(type=type, parent=parent, fragment=fragment, name=name, page_size=page_size))
+        return list(self.select(
+            expression=expression,
+            type=type,
+            parent=parent,
+            fragment=fragment,
+            name=name,
+            owner=owner,
+            query=query,
+            text=text,
+            ids=ids,
+            page_size=page_size,
+            page_number=page_number))
 
     def create(self, *groups):
         """Batch create a collection of groups and entire group trees.
 
-        :param groups:  collection of DeviceGroup instances; each can
-            define children as needed.
+        Args:
+            *groups (DeviceGroup):  collection of DeviceGroup instances;
+                each can define children as needed.
         """
         super()._create(DeviceGroup.to_json, *groups)
 
-    def assign_children(self, root_id, *child_ids):
+    def assign_children(self, root_id: str, *child_ids: str):
         """Link child groups to this device group.
 
         Args:
-            root_id (str|int): ID of the root device group
-            child_ids (*str|int): ID of the child device groups
+            root_id (str): ID of the root device group.
+            *child_ids (str): Collection of the child device group ID.
         """
         # adding multiple references at once is not (yet) supported
         # refs = {'references': [InventoryUtil.build_managed_object_reference(id) for id in child_ids]}
@@ -424,36 +721,36 @@ class DeviceGroupInventory(Inventory):
         """Unlink child groups from this device group.
 
         Args:
-            root_id (str|int): ID of the root device group
-            child_ids (*str|int): ID of the child device groups
+            root_id (str): ID of the root device group.
+            *child_ids (str): Collection of the child device group ID.
         """
         refs = {'references': [ManagedObjectUtil.build_managed_object_reference(i) for i in child_ids]}
         self.c8y.delete(self.build_object_path(root_id) + '/childAssets', json=refs)
 
-    def delete(self, *groups: DeviceGroup | str | int):
+    def delete(self, *groups: DeviceGroup | str):
         """Delete one or more single device groups within the database.
 
         The child groups (if there are any) are left dangling. This is
-        equivalent to using the `cascase=false` parameter in the
+        equivalent to using the `cascade=false` parameter in the
         Cumulocity REST API.
 
         Args:
-            groups:  Objects resp. their ID within the database
+            *groups (str|DeviceGroup):  Collection of objects (or ID).
         """
         self._delete(False, *groups)
 
-    def delete_trees(self, *groups: DeviceGroup | str | int):
+    def delete_trees(self, *groups: DeviceGroup | str):
         """Delete one or more device groups trees within the database.
 
         This is equivalent to using the `cascade=true` parameter in the
         Cumulocity REST API.
 
         Args:
-            groups:  Objects resp. their ID within the database
+            *groups (str|DeviceGroup):  Collection of objects (or ID).
         """
         self._delete(False, *groups)
 
-    def _delete(self, cascade: bool, *objects: DeviceGroup | str | int):
+    def _delete(self, cascade: bool, *objects: DeviceGroup | str):
         try:
             object_ids = [o.id for o in objects]  # noqa (id)
         except AttributeError:
