@@ -7,7 +7,7 @@
 from __future__ import annotations
 
 import dataclasses
-import datetime
+from datetime import datetime, timedelta
 from typing import Type, List, Generator, Sequence
 from urllib.parse import urlencode
 
@@ -122,17 +122,17 @@ class Measurement(ComplexObject):
     # _accept
     # _not_updatable
 
-    def __init__(self, c8y=None, type=None, source=None, time=None, **kwargs):  # noqa (type)
+    def __init__(self, c8y=None, type=None, source=None, time: str|datetime = None, **kwargs):  # noqa (type)
         """ Create a new Measurement object.
 
-        Params:
-            c8y(CumulocityRestApi):  Cumulocity connection reference; needs
+        Args:
+            c8y (CumulocityRestApi)  Cumulocity connection reference; needs
                 to be set for direct manipulation (create, delete)
-            type(str):  Measurement type
-            source(str):  Device ID which this measurement is for
+            type (str)  Measurement type
+            source (str)  Device ID which this measurement is for
             time(str|datetime):  Datetime string or Python datetime object. A
                 given datetime string needs to be in standard ISO format incl.
-                timezone: YYYY-MM-DD'T'HH:MM:SS.SSSZ as it is retured by the
+                timezone: YYYY-MM-DD'T'HH:MM:SS.SSSZ as it is returned by the
                 Cumulocity REST API. A given datetime object needs to be
                 timezone aware. For manual construction it is recommended to
                 specify a datetime object as the formatting of a timestring
@@ -149,7 +149,7 @@ class Measurement(ComplexObject):
         # The time can either be set as string (e.g. when read from JSON) or
         # as a datetime object. It will be converted to string immediately
         # as there is no scenario where a manually created object won't be
-        # written to Cumulocity anyways
+        # written to Cumulocity anyway
         self.time = _DateUtil.ensure_timestring(time)
 
     @classmethod
@@ -159,8 +159,8 @@ class Measurement(ComplexObject):
         The JSON is assumed to be in the format as it is used by the
         Cumulocity REST API.
 
-        Params:
-            measurement_json(dict):  JSON object (nested dictionary)
+        Args:
+            json (dict)  JSON object (nested dictionary)
                 representing a measurement within Cumulocity
 
         Returns:
@@ -216,7 +216,7 @@ class Measurement(ComplexObject):
 
     def update(self) -> Measurement:
         """Not implemented for Measurements."""
-        raise NotImplementedError('Measurements cannot be updated within Cumuylocity.')
+        raise NotImplementedError('Measurements cannot be updated within Cumulocity.')
 
     def delete(self):
         """Delete the Measurement within the database."""
@@ -249,7 +249,7 @@ class Series(dict):
     @property
     def truncated(self):
         """Whether the result was truncated
-        (i.e. the query returned more that 5000 values)."""
+        (i.e. the query returned more than 5000 values)."""
         return self['truncated']
 
     @property
@@ -261,13 +261,13 @@ class Series(dict):
                 timestamps: bool | str = None) -> List | List[tuple]:
         """Collect series results.
 
-        Params:
+        Args:
             series (str|Sequence[str]):  Which series' values to collect. If
                 multiple series are collected each element in the result will
                 be a tuple. If omitted, all available series are collected.
             value (str):  Which value (min/max) to collect. If omitted, both
                 values will be collected, grouped as 2-tuples.
-            timestamp (bool|str):  Whether each element in the result list will
+            timestamps (bool|str):  Whether each element in the result list will
                 be prepended with the corresponding timestamp. If True, the
                 timestamp string will be included; Use 'datetime' or 'epoch' to
                 parse the timestamp string.
@@ -286,9 +286,9 @@ class Series(dict):
         def parse_timestamp(t):
             """Parse timestamps."""
             if timestamps == 'datetime':
-                return datetime.datetime.fromisoformat(t)
+                return datetime.fromisoformat(t)
             if timestamps == 'epoch':
-                return datetime.datetime.fromisoformat(t).timestamp()
+                return datetime.fromisoformat(t).timestamp()
             return t
 
         # use all series if no series provided
@@ -388,23 +388,35 @@ class Measurements(CumulocityResource):
     def get(self, measurement_id: str | int) -> Measurement:
         """ Read a specific measurement from the database.
 
-        params:
+        Args:
             measurement_id (str|int):  database ID of a measurement
 
         Returns:
             Measurement object
 
         Raises:
-            KeyError: If the ID cannot be resolved.
+            KeyError:  if the ID cannot be resolved.
         """
         measurement = Measurement.from_json(self._get_object(measurement_id))
         measurement.c8y = self.c8y  # inject c8y connection into instance
         return measurement
 
-    def select(self, type=None, source=None,  # noqa (type)
-               fragment=None, value=None, series=None,
-               before=None, after=None, min_age=None, max_age=None, reverse=False,
-               limit=None, page_size=1000) -> Generator[Measurement]:
+    def select(
+            self,
+            type: str = None,
+            source: str | int = None,
+            fragment: str = None,
+            value: str = None,
+            series: str = None,
+            before: str | datetime = None,
+            after: str | datetime = None,
+            min_age: timedelta = None,
+            max_age: timedelta = None,
+            reverse: bool = None,
+            limit: int = None,
+            page_size: int = 1000,
+            page_number: int = None
+    ) -> Generator[Measurement]:
         """ Query the database for measurements and iterate over the results.
 
         This function is implemented in a lazy fashion - results will only be
@@ -414,11 +426,11 @@ class Measurements(CumulocityResource):
         to objects which meet the filters specification.  Filters can be
         combined (within reason).
 
-        Params:
+        Args:
             type (str):  Alarm type
-            source (str):  Database ID of a source device
+            source (str|int):  Database ID of a source device
             fragment (str):  Name of a present custom/standard fragment
-            value (str):  Name/type of a present value fragment
+            value (str):  Type/Name of a present value fragment
             series (str):  Name of a present series within a value fragment
             before (datetime|str):  Datetime object or ISO date/time string.
                 Only measurements assigned to a time before this date are
@@ -436,6 +448,8 @@ class Measurements(CumulocityResource):
             page_size (int):  Define the number of measurements which are
                 read (and parsed in one chunk). This is a performance
                 related setting.
+            page_number (int): Pull a specific page; this effectively disables
+                automatic follow-up page retrieval.
 
         Returns:
             Generator[Measurement]: Iterable of matching Measurement objects
@@ -444,12 +458,24 @@ class Measurements(CumulocityResource):
                                             valueFragmentType=value, valueFragmentSeries=series,
                                             before=before, after=after, min_age=min_age, max_age=max_age,
                                             reverse=reverse, page_size=page_size)
-        return super()._iterate(base_query, limit, Measurement.from_json)
+        return super()._iterate(base_query, page_number, limit, Measurement.from_json)
 
-    def get_all(self, type=None, source=None,  # noqa (type)
-                fragment=None,  value=None, series=None,
-                before=None, after=None, min_age=None, max_age=None, reverse=False,
-                limit=None, page_size=1000) -> List[Measurement]:
+    def get_all(
+            self,
+            type: str = None,
+            source: str | int = None,
+            fragment: str = None,
+            value: str = None,
+            series: str = None,
+            before: str | datetime = None,
+            after: str | datetime = None,
+            min_age: timedelta = None,
+            max_age: timedelta = None,
+            reverse: bool = None,
+            limit: int = None,
+            page_size: int = 1000,
+            page_number: int = None
+    ) -> List[Measurement]:
         """ Query the database for measurements and return the results
         as list.
 
@@ -459,13 +485,31 @@ class Measurements(CumulocityResource):
         Returns:
             List of matching Measurement objects
         """
-        return list(self.select(type=type, source=source,
-                                fragment=fragment, value=value, series=series,
-                                before=before, after=after, min_age=min_age, max_age=max_age,
-                                reverse=reverse, limit=limit, page_size=page_size))
+        return list(self.select(
+            type=type,
+            source=source,
+            fragment=fragment,
+            value=value,
+            series=series,
+            before=before,
+            after=after,
+            min_age=min_age,
+            max_age=max_age,
+            reverse=reverse,
+            limit=limit,
+            page_size=page_size,
+            page_number=page_number))
 
-    def get_last(self, type=None, source=None, fragment=None, value=None, series=None,  # noqa (type)
-                 before=None, min_age=None) -> Measurement:
+    def get_last(
+            self,
+            type: str = None,
+            source: str | int = None,
+            fragment: str = None,
+            value: str = None,
+            series: str = None,
+            before: str | datetime = None,
+            min_age: timedelta = None
+    ) -> Measurement:
         """ Query the database and return the last matching measurement.
 
         This function is a special variant of the select function. Only
@@ -479,18 +523,34 @@ class Measurements(CumulocityResource):
         after = None
         if not before and not min_age:
             after = '1970-01-01'
-        base_query = self._build_base_query(type=type, source=source,
-                                            fragment=fragment, value=value, series=series, after=after,
-                                            before=before, min_age=min_age, reverse=True, page_size=1)
-        m = Measurement.from_json(self._get_page(base_query, "1")[0])
+        base_query = self._build_base_query(
+            type=type,
+            source=source,
+            fragment=fragment,
+            value=value,
+            series=series,
+            after=after,
+            before=before,
+            min_age=min_age,
+            reverse=True,
+            page_size=1)
+        m = Measurement.from_json(self._get_page(base_query, page_number=1)[0])
         m.c8y = self.c8y  # inject c8y connection into instance
         return m
 
-    def get_series(self, source: str = None, aggregation: str = None, series: str | Sequence[str] = None,
-                   before=None, after=None, min_age=None, max_age=None, reverse=False) -> Series:
+    def get_series(
+            self,
+            source: str = None,
+            aggregation: str = None,
+            series: str | Sequence[str] = None,
+            before: str | datetime = None,
+            after: str | datetime = None,
+            min_age: timedelta = None,
+            max_age: timedelta = None,
+            reverse=False) -> Series:
         """Query the database for a list of series and their values.
 
-        Params:
+        Args:
             source (str):  Database ID of a source device
             aggregation (str):  Aggregation type
             series (str|Sequence[str]):  Series' to query
@@ -513,9 +573,15 @@ class Measurements(CumulocityResource):
 
         See also: https://cumulocity.com/api/core/#operation/getMeasurementSeriesResource
         """
-        params = self._prepare_query_params(source=source, aggregationType=aggregation,
-                                            before=before, after=after, min_age=min_age, max_age=max_age,
-                                            reverse=reverse)
+        params = self._prepare_query_params(
+            source=source,
+            # this is a non-mapped parameter
+            aggregationType=aggregation,
+            before=before,
+            after=after,
+            min_age=min_age,
+            max_age=max_age,
+            reverse=reverse)
         # The 'series' parameter has to be added manually; because it
         # may be a list and because 'series' is by default converted to
         # the 'valueFragmentSeries' parameter
@@ -527,30 +593,44 @@ class Measurements(CumulocityResource):
         series_json = self.c8y.get(url)
         return Series(series_json)
 
-    def delete_by(self, type=None, source=None,  # noqa (type)
-                fragment=None, value=None, series=None,
-                before=None, after=None, min_age=None, max_age=None):
+    def delete_by(
+            self,
+            type: str = None,
+            source: str | int = None,
+            fragment: str = None,
+            value: str = None,
+            series: str = None,
+            before: str | datetime = None,
+            after: str | datetime = None,
+            min_age: timedelta = None,
+            max_age: timedelta = None,
+        ):
         """ Query the database and delete matching measurements.
 
-         All parameters are considered to be filters, limiting the result set
+        All parameters are considered to be filters, limiting the result set
         to objects which meet the filters specification.  Filters can be
         combined (within reason).
 
-        Params: See 'select' function
+        Args: See 'select' function
         """
-        base_query = self._build_base_query(type=type, source=source,
-                                            fragment=fragment, value=value, series=series,
-                                            before=before, after=after, min_age=min_age, max_age=max_age)
-        # remove &page_number= from the end
-        query = base_query[:base_query.rindex('&')]
-        self.c8y.delete(query)
+        base_query = self._build_base_query(
+            type=type,
+            source=source,
+            fragment=fragment,
+            value=value,
+            series=series,
+            before=before,
+            after=after,
+            min_age=min_age,
+            ax_age=max_age)
+        self.c8y.delete(base_query)
 
     # delete function is defined in super class
 
     def create(self, *measurements):
         """ Bulk create a collection of measurements within the database.
 
-        Params:
-            measurements (Iterable): Iterable collection of Measurement objects.
+        Args:
+            *measurements (Measurement): Collection of Measurement objects.
         """
         self._create_bulk(Measurement.to_json, 'measurements', self.c8y.CONTENT_MEASUREMENT_COLLECTION, *measurements)
