@@ -70,7 +70,7 @@ class Alarm(ComplexObject):
             status (str):  Alarm status
             severity (str):  Alarm severity
             kwargs:  Additional arguments are treated as custom fragments
-    """
+        """
         super().__init__(c8y=c8y, **kwargs)
         self.type = type
         self.source = source
@@ -88,6 +88,7 @@ class Alarm(ComplexObject):
         self.time = _DateUtil.ensure_timestring(time)
 
     text = SimpleObject.UpdatableProperty('_u_text')
+    """Updatable property"""
     status = SimpleObject.UpdatableProperty('_u_status')
     severity = SimpleObject.UpdatableProperty('_u_severity')
 
@@ -123,7 +124,7 @@ class Alarm(ComplexObject):
         """Convert the first occurrence time to a Python datetime object.
 
         Returns:
-            Standard Python datetime object for the first occurance time.
+            Standard Python datetime object for the first occurrence time.
         """
         return super()._to_datetime(self.first_occurrence_time)
 
@@ -204,7 +205,7 @@ class Alarm(ComplexObject):
         must be defined for this to function. This is always the case if
         the instance was built by the API.
 
-        See also functions Alarms.delete and Alarms.delete_by
+        See also functions `Alarms.delete` and `Alarms.delete_by`
         """
         self._assert_c8y()
         if not self.type:
@@ -247,7 +248,8 @@ class Alarms(CumulocityResource):
                created_before: str | datetime = None, created_after: str | datetime = None,
                updated_before: str | datetime = None, updated_after: str | datetime = None,
                min_age: timedelta = None, max_age: timedelta = None,
-               reverse: bool = False, limit: int = None, page_size: int = 1000) -> Generator[Alarm]:
+               reverse: bool = False, limit: int = None,
+               page_size: int = 1000, page_number: int = None) -> Generator[Alarm]:
         """Query the database for alarms and iterate over the results.
 
         This function is implemented in a lazy fashion - results will only be
@@ -283,6 +285,8 @@ class Alarms(CumulocityResource):
             limit (int): Limit the number of results to this number.
             page_size (int): Define the number of alarms which are read (and
                 parsed in one chunk). This is a performance related setting.
+            page_number (int): Pull a specific page; this effectively disables
+                automatic follow-up page retrieval.
 
         Returns:
             Generator of Alarm objects
@@ -294,7 +298,7 @@ class Alarms(CumulocityResource):
                                             updated_before=updated_before, updated_after=updated_after,
                                             min_age=min_age, max_age=max_age,
                                             reverse=reverse, page_size=page_size)
-        return super()._iterate(base_query, limit, Alarm.from_json)
+        return super()._iterate(base_query, page_number, limit, Alarm.from_json)
 
     def get_all(self, type: str = None, source: str = None, fragment: str = None, # noqa (type)
                 status: str = None, severity: str = None, resolved: str = None,
@@ -302,7 +306,8 @@ class Alarms(CumulocityResource):
                 created_before: str | datetime = None, created_after: str | datetime = None,
                 updated_before: str | datetime = None, updated_after: str | datetime = None,
                 min_age: timedelta = None, max_age: timedelta = None,
-                reverse: bool = False, limit: int = None, page_size: int = 1000) -> List[Alarm]:
+                reverse: bool = False, limit: int = None,
+                page_size: int = 1000, page_number: int = None) -> List[Alarm]:
         """Query the database for alarms and return the results as list.
 
         This function is a greedy version of the select function. All
@@ -319,7 +324,7 @@ class Alarms(CumulocityResource):
                                 created_before=created_before, created_after=created_after,
                                 updated_before=updated_before, updated_after=updated_after,
                                 min_age=min_age, max_age=max_age, reverse=reverse,
-                                limit=limit, page_size=page_size))
+                                limit=limit, page_size=page_size, page_number=page_number))
 
     def count(self, type: str = None, source: str = None, fragment: str = None, # noqa (type)
                 status: str = None, severity: str = None, resolved: str = None,
@@ -354,7 +359,7 @@ class Alarms(CumulocityResource):
         """Create alarm objects within the database.
 
         Args:
-            alarms (*Alarm): Collection of Alarm instances
+            *alarms (Alarm): Collection of Alarm instances
         """
         super()._create(Alarm.to_full_json, *alarms)
 
@@ -362,16 +367,17 @@ class Alarms(CumulocityResource):
         """Write changes to the database.
 
         Args:
-            alarms (*Alarm): Collection of Alarm instances
+            *alarms (Alarm): Collection of Alarm instances
         """
         super()._update(Alarm.to_diff_json, *alarms)
 
-    def apply_to(self, alarm, *alarm_ids):
+    def apply_to(self, alarm: Alarm|dict, *alarm_ids: str):
         """Apply changes made to a single instance to other objects in the database.
 
         Args:
-            alarm (Alarm): Object serving as model for the update
-            alarm_ids (*str): A collection of database IDS of alarms
+            alarm (Alarm|dict): Object serving as model for the update or
+                simply a dictionary representing the diff JSON.
+            *alarm_ids (str): A collection of database IDS of alarms
         """
         super()._apply_to(Alarm.to_full_json, alarm, *alarm_ids)
 
@@ -411,7 +417,7 @@ class Alarms(CumulocityResource):
         operation to function.
 
         Args:
-            alarms (*Alarm): Collection of Alarm instances.
+            *alarms (Alarm): Collection of Alarm instances.
         """
         for a in alarms:
             a.delete()
@@ -444,6 +450,4 @@ class Alarms(CumulocityResource):
         base_query = self._build_base_query(type=type, source=source, fragment=fragment,
                                             status=status, severity=severity, resolved=resolved,
                                             before=before, after=after, min_age=min_age, max_age=max_age)
-        # remove &page_number= from the end
-        query = base_query[:base_query.rindex('&')]
-        self.c8y.delete(query)
+        self.c8y.delete(base_query)
