@@ -45,12 +45,13 @@ class Subscription(SimpleObject):
     _parser = SimpleObjectParser({
             'name': 'subscription',
             'context': 'context',
+            'non_persistent': 'nonPersistent',
             'fragments': 'fragmentsToCopy'})
     _accept = 'application/vnd.com.nsn.cumulocity.subscription+json'
 
     def __init__(self, c8y: CumulocityRestApi = None, name: str = None, context: str = None, source_id: str = None,
                  api_filter: List[str] = None, type_filter: str = None,
-                 fragments: List[str] = None):
+                 fragments: List[str] = None, non_persistent: bool = None):
         """ Create a new Subscription instance.
 
         Args:
@@ -61,6 +62,7 @@ class Subscription(SimpleObject):
             source_id (str):  Managed object ID the subscription is for.
             api_filter (List[str]):  List of APIs/resources to subscribe for.
             type_filter (str):  Object type the subscription is for.
+            non_persistent (bool):  Whether the subscription is non-persistent.
 
         Returns:
             Subscription instance
@@ -69,6 +71,7 @@ class Subscription(SimpleObject):
         self.name = name
         self.context = context
         self.source_id = source_id
+        self.non_persistent = non_persistent
         self.api_filter = api_filter
         self.type_filter = type_filter
         self.fragments = fragments
@@ -87,11 +90,7 @@ class Subscription(SimpleObject):
     def from_json(cls, json: dict) -> Subscription:
         """Create a Subscription instance from Cumulocity JSON format.
 
-        Caveat: this function is primarily for internal use and does not
-        return a full representation of the JSON. It is used for object
-        creation and update within Cumulocity.
-
-        Params:
+        Args:
             json (dict): The JSON to parse.
 
         Returns:
@@ -113,7 +112,7 @@ class Subscription(SimpleObject):
             A fresh Subscription instance representing the created
             subscription within the database.
 
-        See also function Subscriptions.create which doesn't parse the result.
+        See also function `Subscriptions.create` which doesn't parse the result.
         """
         return self._create()
 
@@ -141,14 +140,14 @@ class Subscriptions(CumulocityResource):
              A Subscription instance
 
         Raises:
-            KeyError if the given ID is not defined within the database
+            KeyError:  if the given ID is not defined within the database
         """
         subscription = Subscription.from_json(super()._get_object(subscription_id))
         subscription.c8y = self.c8y  # inject c8y connection into instance
         return subscription
 
-    def select(self, context: str = None, source: str = None,
-               limit: int = None, page_size: int = 1000) -> Generator[Subscription]:
+    def select(self, context: str = None, source: str = None, subscription: str = None,
+               limit: int = None, page_size: int = 1000, page_number: int = None) -> Generator[Subscription]:
         """ Query the database for subscriptions and iterate over the
         results.
 
@@ -162,18 +161,22 @@ class Subscriptions(CumulocityResource):
         Args:
             context (str):  Subscription context.
             source (str):  Managed object ID the subscription is for.
+            subscription (str): The subscription name.
             limit (int): Limit the number of results to this number.
             page_size (int): Define the number of objects which are read (and
                 parsed in one chunk). This is a performance related setting.
+            page_number (int): Pull a specific page; this effectively disables
+                automatic follow-up page retrieval.
 
         Returns:
             Generator for Subscription instances
         """
-        base_query = self._build_base_query(context=context, source=source, page_size=page_size)
-        return super()._iterate(base_query, limit, Subscription.from_json)
+        base_query = self._build_base_query(context=context, source=source,
+                                            subscription=subscription, page_size=page_size)
+        return super()._iterate(base_query, page_number, limit, Subscription.from_json)
 
-    def get_all(self, context: str = None, source: str = None,
-                limit: int = None, page_size: int = 1000) -> List[Subscription]:
+    def get_all(self, context: str = None, source: str = None, subscription: str = None,
+                limit: int = None, page_size: int = 1000, page_number: int = None) -> List[Subscription]:
         """ Query the database for subscriptions and return the results
         as list.
 
@@ -183,13 +186,14 @@ class Subscriptions(CumulocityResource):
         Returns:
             List of Subscription instances.
         """
-        return list(self.select(context=context, source=source, limit=limit, page_size=page_size))
+        return list(self.select(context=context, source=source, subscription=subscription, limit=limit,
+                                page_size=page_size, page_number=page_number))
 
     def create(self, *subscriptions: Subscription) -> None:
         """ Create subscriptions within the database.
 
         Args:
-            subscriptions (*TenantOption):  Collection of Subscription instances
+            *subscriptions (TenantOption):  Collection of Subscription instances
         """
         super()._create(Subscription.to_full_json, *subscriptions)
 
@@ -223,23 +227,31 @@ class Tokens(CumulocityResource):
         super().__init__(c8y, '/notification2')
         self.host = urllib.parse.urlparse(c8y.base_url).netloc
 
-    def generate(self, subscription: str, expires: int = 60, subscriber: str = None) -> str:
+    def generate(self, subscription: str, expires: int = 60, subscriber: str = None,
+                 signed: bool = None, shared: bool = None, non_persistent: bool = None) -> str:
         """Generate a new access token.
 
         Args:
-            subscription (str): Subscription name
-            expires (int):  Expiration time in minutes
-            subscriber (str):  Subscriber Id (name)
+            subscription (str): Subscription name.
+            expires (int):  Expiration time in minutes.
+            subscriber (str):  Subscriber ID (name). A UUID based default will be used if None.
+            signed (bool):  Whether the token should be signed.
+            shared (bool):  Whether the token is used to create a shared consumer.
+            non_persistent (bool): Whether the token refers to the non-persistent variant of the named subscription.
 
         Returns:
             JWT access token as string.
         """
-        td_json = self._build_token_definition(subscription, expires, subscriber)
+        td_json = self._build_token_definition(subscription, expires, subscriber, signed, shared, non_persistent)
         token_json = self.c8y.post(self.resource + '/token', td_json)
         return token_json['token']
 
-    def renew(self, token: str):
-        """Renew a token."""
+    def renew(self, token: str) -> str:
+        """Renew a token.
+
+        Args:
+            token:  Currently valid token to be renewed.
+        """
 
     def unsubscribe(self, token: str):
         """Invalidate a token and unsubscribe a subscriber.
@@ -251,21 +263,31 @@ class Tokens(CumulocityResource):
         if not result_json['result'] == 'DONE':
             raise RuntimeError(f"Unexpected response: {js.dumps(result_json)}")
 
-    def build_websocket_uri(self, token: str):
+    def build_websocket_uri(self, token: str, consumer: str = None):
         """Build websocket access URL.
 
         Args:
             token (str):  Subscriber access token
+            consumer (str): Optional consumer ID (to allow 'sticky' connections after interrupt).
 
         Returns:
              A websocket (ws(s)://) URL to access the subscriber channel.
         """
         protocol = 'wss' if self.c8y.is_tls else 'ws'
-        return f'{protocol}://{self.host}/notification2/consumer/?token={token}'
+        consumer_param = f'&consumer={consumer}' if consumer else ''
+        return f'{protocol}://{self.host}/notification2/consumer/?token={token}{consumer_param}'
 
-    def _build_token_definition(self, subscription: str, expires: int, subscriber: str = None):
-        return {
+    def _build_token_definition(self, subscription: str, expires: int, subscriber: str = None,
+                                signed = None, shared = None, non_persistent = None):
+        json =  {
             'subscriber': subscriber or self._default_subscriber,
             'subscription' : subscription,
-            'expiresInMinutes' : expires
+            'expiresInMinutes' : expires,
         }
+        if signed is not None:
+            json['signed'] = signed
+        if shared is not None:
+            json['shared'] = shared
+        if non_persistent is not None:
+            json['nonPersistent'] = non_persistent
+        return json
