@@ -9,7 +9,7 @@ import pytest
 from c8y_api import CumulocityApi
 from c8y_api.model.identity import ExternalId
 
-from tests import RandomNameGenerator
+from util.testing_util import RandomNameGenerator
 
 
 def test_CRUD(live_c8y: CumulocityApi, sample_device):
@@ -52,4 +52,3 @@ def test_CRUD(live_c8y: CumulocityApi, sample_device):
 
     with pytest.raises(KeyError):
         live_c8y.identity.get(id_ref2, id_type)
-
