@@ -8,7 +8,7 @@
 from c8y_api import CumulocityApi
 from c8y_api.model import Operation
 
-from tests import RandomNameGenerator
+from util.testing_util import RandomNameGenerator
 
 
 def test_CRUD(live_c8y: CumulocityApi, sample_device):
