@@ -15,7 +15,7 @@ import pyotp
 from c8y_api import CumulocityApi
 from c8y_api.model import User
 
-from tests import RandomNameGenerator
+from util.testing_util import RandomNameGenerator
 
 
 def test_CRUD(live_c8y: CumulocityApi):  # noqa (case)
