@@ -4,15 +4,14 @@
 # Use, reproduction, transfer, publication or disclosure is prohibited except
 # as specifically provided for in your License Agreement with Software AG.
 
-# pylint: disable=redefined-outer-name
+# pylint: disable=redefined-outer-name, protected-access
 
 import random
 import pytest
 
-from tests.utils import RandomNameGenerator
+from c8y_api.model._parser import SimpleObjectParser
 
-# pylint: disable=protected-access
-from c8y_api.model._parser import SimpleObjectParser  # noqa
+from util.testing_util import RandomNameGenerator
 
 
 @pytest.fixture(scope='function')
@@ -100,7 +99,7 @@ def test_to_json_simple(simple_object_and_mapping):
 
 def test_to_json_simple_includes(simple_object_and_mapping):
     """Verify that simple JSON rendering works using the to_json method
-    of a parser instance when an include is is specified."""
+    of a parser instance when an include is specified."""
 
     obj, mapping = simple_object_and_mapping
     parser = SimpleObjectParser(mapping)
