@@ -14,7 +14,8 @@ import pytest
 
 from c8y_api import CumulocityRestApi
 from c8y_api.model.notification2 import Subscription
-from utils import isolate_last_call_arg
+
+from tests.utils import isolate_last_call_arg
 
 
 def fix_sample_jsons() -> List[dict]:
@@ -27,12 +28,14 @@ def fix_sample_jsons() -> List[dict]:
 
 @pytest.mark.parametrize('sample_json', fix_sample_jsons())
 def test_parsing(sample_json):
-    """Verify that parsing a Subscription from JSON works."""
+    """Verify that parsing a Subscription from JSON works as expected."""
     subscription = Subscription.from_json(sample_json)
 
     assert subscription.name == sample_json['subscription']
     assert subscription.context == sample_json['context']
     assert subscription.source_id == sample_json['source']['id']
+    if 'nonPersistent' in sample_json:
+        assert subscription.non_persistent == sample_json['nonPersistent']
     if 'fragmentsToCopy' in sample_json:
         assert subscription.fragments == sample_json['fragmentsToCopy']
     if 'subscriptionFilter' in sample_json:
@@ -44,9 +47,11 @@ def test_parsing(sample_json):
 
 
 def test_formatting():
-    subscription = Subscription(name='name', source_id='source_id', context=Subscription.Context.TENANT)
-    subscription_json = subscription.to_full_json()
+    """Verify that JSON formatting works as expected."""
 
+    subscription = Subscription(name='name', source_id='source_id', context=Subscription.Context.TENANT)
+
+    subscription_json = subscription.to_full_json()
     assert subscription_json['subscription'] == 'name'
     assert subscription_json['context'] == 'tenant'
     assert subscription_json['source']['id'] == 'source_id'
@@ -62,15 +67,18 @@ def test_formatting():
     assert len(subscription_json) == 5
 
     subscription.api_filter = ['a1', 'a2']
+    subscription.non_persistent = True
     subscription_json = subscription.to_full_json()
     assert subscription_json['fragmentsToCopy'] == subscription.fragments
     assert subscription_json['subscriptionFilter']['typeFilter'] == subscription.type_filter
     assert subscription_json['subscriptionFilter']['apis'] == subscription.api_filter
+    assert subscription_json['nonPersistent']
     # expect no other entries
-    assert len(subscription_json) == 5
+    assert len(subscription_json) == 6
 
 
 def test_create():
+    """Verify that the object creation works as expected (JSON & URLs)."""
     c8y: CumulocityRestApi = Mock()
     # the post function can just return any valid JSON
     sample_json = fix_sample_jsons()[0]
@@ -85,7 +93,7 @@ def test_create():
     assert subscription.to_json.call_count == 1
     # 2) the given resource path should be correct
     resource = isolate_last_call_arg(c8y.post, 'resource', 0)
-    assert resource == f'/notification2/subscriptions'
+    assert resource == '/notification2/subscriptions'
     # 3) the given payload should match what to_json returned
     payload = isolate_last_call_arg(c8y.post, 'json', 1)
     assert set(payload.keys()) == {'expected'}
@@ -96,6 +104,7 @@ def test_create():
 
 
 def test_delete():
+    """Verify that the object deletion works as expected."""
     c8y: CumulocityRestApi = Mock()
     c8y.delete = Mock()
     subscription = Subscription(c8y=c8y, name='name', source_id='source_id', context=Subscription.Context.TENANT)
@@ -107,4 +116,3 @@ def test_delete():
     # 2) the given resource path should be correct
     resource = isolate_last_call_arg(c8y.delete, 'resource', 0)
     assert resource == f'/notification2/subscriptions/{subscription.id}'
-
